# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Цифровой двойник цеха — точка входа",
    "version": "1.0.0",
    "description": "Единая документированная точка входа двойника. По этим адресам ходят имитаторы 1С в прототипе и пойдут настоящие системы завода.\n\n- Все события проверяются схемой; ошибки сохраняются и видны на экране «Источники данных».\n- Повтор события с тем же `eventId` не создаёт дубль (идемпотентность).\n- Время — ISO 8601 с часовым поясом, завод работает по UTC+5.\n- Контроллеры и видеоаналитика подключаются по MQTT, см. «Документация MQTT (AsyncAPI)».\n\nТипичная интеграция с 1С:MES/QLS/WMS — HTTP-сервисы или OData; конкретная выгрузка настраивается со специалистами завода."
  },
  "servers": [
    {
      "url": "/",
      "description": "Шлюз двойника"
    }
  ],
...
  "paths": {
    "/api/v1/events": {
      "post": {
        "tags": [
          "1С и мастер"
        ],
        "summary": "Пакет событий из 1С:MES / QLS / WMS",
        "description": "Массив канонических событий. Каждое проверяется отдельно: корректные принимаются, ошибочные возвращаются с причиной.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventBatch"
              },
              "examples": {
                "Проход поста и несоответствие": {
                  "value": [
                    {
                      "eventId": "mes-20261007-0001240",
                      "source": "mes",
                      "ts": "2026-10-07T13:40:00+05:00",
                      "area": "weld",
                      "equipmentId": "ABB-04",
                      "vin": "KZACB2S5XTK004828",
                      "type": "post_passed",
                      "payload": {
                        "model": "cobalt",
                        "post": "WELD-4"
                      }
                    },
                    {
                      "eventId": "qls-20261007-000078",
                      "source": "qls",
                      "ts": "2026-10-07T13:40:30+05:00",
                      "area": "paint",
                      "vin": "KZACN1S13TK004815",
                      "type": "nonconformity",
                      "payload": {
                        "checkpoint": "CP-PAINT",
                        "defect": "paint_dirt",
                        "decision": "repaint"
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Итог приёма",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResult"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации (сообщения по-русски)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plan": {
      "post": {
        "tags": [
          "1С и мастер"
        ],
        "summary": "План на месяц по моделям (1С:ERP)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              },
              "examples": {
                "План октября": {
                  "value": {
                    "month": "2026-10",
                    "target": 5500,
                    "models": [
                      {
                        "model": "onix",
                        "qty": 2500
                      },
                      {
                        "model": "cobalt",
                        "qty": 1800
                      },
                      {
                        "model": "j7",
                        "qty": 500
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Итог приёма",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResult"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации (сообщения по-русски)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shift-reports": {
      "post": {
        "tags": [
          "1С и мастер"
        ],
        "summary": "Сменный отчёт в формате выданной таблицы",
        "description": "Дата, Линия, План, Факт, Время работы (ч), Загрузка (%). Дата — ДД.ММ.ГГГГ или ГГГГ-ММ-ДД.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftReportRequest"
              },
              "examples": {
                "Данные организаторов": {
                  "value": [
                    {
                      "date": "01.10.2026",
                      "line": "Сварка-1",
                      "plan": 120,
                      "fact": 118,
                      "hours": 7.8,
                      "load": 98
                    },
                    {
                      "date": "01.10.2026",
                      "line": "Окраска-1",
                      "plan": 120,
                      "fact": 115,
                      "hours": 7.5,
                      "load": 94
                    },
                    {
                      "date": "01.10.2026",
                      "line": "Сборка-1",
                      "plan": 120,
                      "fact": 121,
                      "hours": 8,
                      "load": 100
                    },
                    {
                      "date": "02.10.2026",
                      "line": "Сварка-1",
                      "plan": 120,
                      "fact": 111,
                      "hours": 7.2,
                      "load": 91
                    },
                    {
                      "date": "02.10.2026",
                      "line": "Окраска-1",
                      "plan": 120,
                      "fact": 116,
                      "hours": 7.7,
                      "load": 96
                    },
                    {
                      "date": "02.10.2026",
                      "line": "Сборка-1",
                      "plan": 120,
                      "fact": 119,
                      "hours": 7.9,
                      "load": 99
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Итог приёма и найденные противоречия",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации (сообщения по-русски)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/downtimes": {
      "post": {
        "tags": [
          "1С и мастер"
        ],
        "summary": "Регистрация простоя (1С:MES или мастер)",
        "description": "Если не указать from — простой начинается сейчас (по часам двойника). Событие сразу появляется на экране руководителя.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DowntimeRequest"
              },
              "examples": {
                "Обрыв цепи на сборке": {
                  "value": {
                    "source": "master",
                    "area": "assembly",
                    "equipmentId": "CONV-03",
                    "reason": "Обрыв цепи",
                    "category": "breakdown",
                    "registeredBy": "Мастер сборки"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Итог приёма",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResult"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации (сообщения по-русски)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quality": {
      "post": {
        "tags": [
          "1С и мастер"
        ],
        "summary": "Итог качества в формате выданной таблицы",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityRequest"
              },
              "examples": {
                "Данные организаторов": {
                  "value": [
                    {
                      "date": "01.10.2026",
                      "area": "Сварка",
                      "produced": 118,
                      "defects": 2,
                      "pct": 1.7
                    },
                    {
                      "date": "01.10.2026",
                      "area": "Окраска",
                      "produced": 115,
                      "defects": 4,
                      "pct": 3.5
                    },
                    {
                      "date": "01.10.2026",
                      "area": "Сборка",
                      "produced": 121,
                      "defects": 1,
                      "pct": 0.8
                    },
                    {
                      "date": "02.10.2026",
                      "area": "Сварка",
                      "produced": 111,
                      "defects": 3,
                      "pct": 2.7
                    },
                    {
                      "date": "02.10.2026",
                      "area": "Окраска",
                      "produced": 116,
                      "defects": 6,
                      "pct": 5.2
                    },
                    {
                      "date": "02.10.2026",
                      "area": "Сборка",
                      "produced": 119,
                      "defects": 2,
                      "pct": 1.7
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Итог приёма и найденные противоречия",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации (сообщения по-русски)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/import/csv": {
      "post": {
        "tags": [
          "Импорт"
        ],
        "summary": "Импорт выданной таблицы файлом CSV",
        "description": "Тело — текст CSV (разделитель «;», «,» или табуляция; десятичная запятая допустима). Тип таблицы определяется по заголовкам: работа линий, простои, качество, план по моделям. В ответе — найденные противоречия.",
        "requestBody": {
          "required": true,
          "content": {
            "text/csv": {
              "schema": {
                "type": "string"
              },
              "example": "Дата;Линия;План;Факт;Время работы, ч;Загрузка, %\n01.10.2026;Сварка-1;120;118;7,8;98"
            }
          }
        },
        "responses": {
          "200": {
            "description": "Итог импорта и найденные противоречия",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации (сообщения по-русски)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/demo/clock": {
      "post": {
        "tags": [
          "Демо"
        ],
        "summary": "Скорость и пауза",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "speed": {
                    "type": "number"
                  },
                  "paused": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Состояние",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/demo/scenario": {
      "post": {
        "tags": [
          "Демо"
        ],
        "summary": "Запустить сценарий",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scenario": {
                    "type": "string"
                  }
                },
                "required": [
                  "scenario"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Состояние",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/demo/stage": {
      "post": {
        "tags": [
          "Демо"
        ],
        "summary": "Ступень внедрения 0/1/2",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "stage": {
                    "$ref": "#/components/schemas/Stage"
                  }
                },
                "required": [
                  "stage"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Состояние",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/demo/simulate-all": {
      "post": {
        "tags": [
          "Демо"
        ],
        "summary": "Имитировать данные со всего оборудования, не дожидаясь подключения",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "on": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "on"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Состояние",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AreaId": {
        "type": "string",
        "pattern": "^[a-z][a-z0-9-]{0,31}$",
        "description": "Код участка из конфигурации завода (GET /api/v1/plant/config). Исходный цех: warehouse — склад комплектующих, weld — сварка, paint — окраска, assembly — сборка, qc — контроль качества (ОТК), finished — склад готовой продукции. Шлюз отклоняет участки, которых нет в конфигурации",
        "examples": [
          "paint"
        ]
      },
      "ModelId": {
        "type": "string",
        "enum": [
          "onix",
          "cobalt",
          "j7"
        ],
        "description": "Модель: onix — Chevrolet Onix, cobalt — Chevrolet Cobalt, j7 — JAC J7"
      },
...
      "Timestamp": {
        "type": "string",
        "format": "date-time",
        "description": "Время в ISO 8601 с часовым поясом",
        "examples": [
          "2026-10-07T13:40:00+05:00"
        ]
      },
      "EventId": {
        "type": "string",
        "minLength": 1,
        "maxLength": 128,
        "description": "Уникальный идентификатор события в системе-источнике. Повтор с тем же eventId не создаёт дубль.",
        "examples": [
          "mes-20261007-000123"
        ]
      },
      "EquipmentId": {
        "type": "string",
        "minLength": 1,
        "maxLength": 64,
        "description": "Код оборудования из конфигурации завода. Исходный цех: роботы сварки ABB-01…ABB-11 и LASER-01 (три линии под модели), FINISH-01, GEO-LAB, PRETREAT, OVEN-ED, SEALER, PRIMER, BOOTH-01, BOOTH-02, OVEN, PAINT-INSP, POLISH, CONV-01…CONV-03, QC-TEST, QC-RAIN, QC-TRACK, QC-LINE"
      },
...
      "DowntimeCategory": {
        "type": "string",
        "enum": [
          "breakdown",
          "no_parts",
          "setup",
          "waiting",
          "planned",
          "quality",
          "other"
        ],
        "description": "breakdown — поломка, no_parts — нет деталей, setup — наладка, waiting — ждём решения, planned — плановое ТО, quality — по качеству, other — другое"
      },
...
      "PlanModelLine": {
        "type": "object",
        "properties": {
          "model": {
            "$ref": "#/components/schemas/ModelId"
          },
          "qty": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "model",
          "qty"
        ]
      },
...
      "CanonicalEvent": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/PostPassedEvent"
          },
          {
            "$ref": "#/components/schemas/NonconformityEvent"
          },
          {
            "$ref": "#/components/schemas/DowntimeRegisteredEvent"
          },
          {
            "$ref": "#/components/schemas/EquipmentStateEvent"
          },
          {
            "$ref": "#/components/schemas/EquipmentCounterEvent"
          },
          {
            "$ref": "#/components/schemas/TelemetryEvent"
          },
          {
            "$ref": "#/components/schemas/CameraDetectionEvent"
          },
          {
            "$ref": "#/components/schemas/StockLevelEvent"
          },
          {
            "$ref": "#/components/schemas/PlanSetEvent"
          },
          {
            "$ref": "#/components/schemas/ShiftReportEvent"
          },
          {
            "$ref": "#/components/schemas/QualitySummaryEvent"
          },
          {
            "$ref": "#/components/schemas/ProductionOrderEvent"
          },
          {
            "$ref": "#/components/schemas/BodyCheckpointEvent"
          },
          {
            "$ref": "#/components/schemas/OperationResultEvent"
          },
          {
            "$ref": "#/components/schemas/VinAssignedEvent"
          }
        ],
        "description": "Каноническое событие двойника; поле type определяет состав payload"
      },
...
      "EventBatch": {
        "minItems": 1,
        "maxItems": 1000,
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/CanonicalEvent"
        },
        "description": "Пакет канонических событий. Каждое событие проверяется отдельно: корректные принимаются, ошибочные возвращаются с причиной."
      },
      "IngestRejected": {
        "type": "object",
        "properties": {
          "index": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "eventId": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          }
        },
        "required": [
          "index",
          "issues"
        ]
      },
      "IngestResult": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "Принято новых событий"
          },
          "duplicates": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "Повторы по eventId — пропущены"
          },
          "rejected": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IngestRejected"
            }
          }
        },
        "required": [
          "accepted",
          "duplicates",
          "rejected"
        ],
        "description": "Итог приёма"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          }
        },
        "required": [
          "error",
          "message"
        ]
      },
      "DowntimeRequest": {
        "type": "object",
        "properties": {
          "eventId": {
            "$ref": "#/components/schemas/EventId"
          },
          "source": {
            "default": "mes",
            "type": "string",
            "enum": [
              "mes",
              "master"
            ]
          },
          "ts": {
            "description": "Время записи; нет — время двойника",
            "$ref": "#/components/schemas/Timestamp"
          },
          "area": {
            "$ref": "#/components/schemas/AreaId"
          },
          "equipmentId": {
            "$ref": "#/components/schemas/EquipmentId"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "category": {
            "$ref": "#/components/schemas/DowntimeCategory"
          },
          "from": {
            "description": "Начало простоя; нет — сейчас",
            "$ref": "#/components/schemas/Timestamp"
          },
          "to": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "registeredBy": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "area",
          "equipmentId",
          "reason",
          "category",
          "registeredBy"
        ],
        "description": "Регистрация простоя из 1С:MES или с телефона мастера"
      },
      "PlanRequest": {
        "type": "object",
        "properties": {
          "eventId": {
            "$ref": "#/components/schemas/EventId"
          },
          "month": {
            "type": "string",
            "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
          },
          "target": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "models": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanModelLine"
            }
          }
        },
        "required": [
          "month",
          "models"
        ],
        "description": "План на месяц по моделям (1С:ERP)"
      },
      "DateLike": {
        "type": "string",
        "pattern": "^(\\d{4}-\\d{2}-\\d{2}|\\d{1,2}\\.\\d{1,2}\\.\\d{4})$"
      },
      "ShiftReportRow": {
        "type": "object",
        "properties": {
          "date": {
            "$ref": "#/components/schemas/DateLike"
          },
          "shift": {
            "anyOf": [
              {
                "type": "number",
                "const": 1
              },
              {
                "type": "number",
                "const": 2
              }
            ]
          },
          "line": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "description": "Линия, как в отчёте: «Сварка-1», «Окраска-1», «Сборка-1»"
          },
          "plan": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "fact": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "hours": {
            "type": "number",
            "minimum": 0,
            "maximum": 24,
            "description": "Время работы, ч"
          },
          "load": {
            "type": "number",
            "minimum": 0,
            "maximum": 200,
            "description": "Загрузка, %"
          }
        },
        "required": [
          "date",
          "line",
          "plan",
          "fact",
          "hours",
          "load"
        ],
        "description": "Строка сменного отчёта в формате выданной таблицы"
      },
      "ShiftReportRequest": {
        "minItems": 1,
        "maxItems": 5000,
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ShiftReportRow"
        }
      },
...
      "QualityRequest": {
        "minItems": 1,
        "maxItems": 5000,
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/QualityRow"
        }
      },
...
      "Stage": {
        "anyOf": [
          {
            "type": "number",
            "const": 0
          },
          {
            "type": "number",
            "const": 1
          },
          {
            "type": "number",
            "const": 2
          }
        ],
        "description": "Ступень внедрения: 0 — только 1С, 1 — + контроллеры и камеры, 2 — + датчики на критичных узлах"
      },
...
    },
...
  }
}